requirements.txt
certifi==2020.6.20
click==7.1.2
decorator==4.4.2
dnspython==1.16.0
dominate==2.4.0
ecdsa==0.13
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycryptodome==3.9.7
python-dateutil==2.8.1
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.3.dev0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
